watchers/pyproject.toml
[project]
name = "watchers"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "watchdog>=6.0.0",
    "google-api-python-client",
    "google-auth-oauthlib",
    "google-auth-httplib2",
]
